+++ src/parsers/problem/PandaOnlineJudgeProblemParser.ts
@@ -16,7 +16,7 @@ export class PandaOnlineJudgeProblemParser extends Parser {
       task.setName(elem.querySelector('.mat-card-title').textContent);
       task.setGroup('Panda Online Judge');
 
-      const blocks = elem.querySelectorAll('pre.sample-item');
+      const blocks = elem.querySelectorAll('pre.sample-box');
       for (let i = 0; i < blocks.length; i += 2) {
         const input = blocks[i].textContent.trim();
         const output = blocks[i + 1].textContent.trim();