+++ src/parsers/problem/TophProblemParser.ts
@@ -13,15 +13,12 @@ export class TophProblemParser extends Parser {
       const elem = htmlToElement(html);
       const task = new TaskBuilder().setUrl(url);
 
-      task.setName(
-        elem.querySelector('.problem-statement-portlet h2').textContent,
-      );
+      task.setName(elem.querySelector('.problem-statement h2').textContent);
 
       task.setGroup('Toph');
 
-      const limitsStr = elem.querySelector(
-        '.problem-statement-portlet div > span',
-      ).textContent;
+      const limitsStr = elem.querySelector('.problem-statement div > span')
+        .textContent;
 
       task.setTimeLimit(
         parseFloat(/Limits: ([0-9.]+)s/.exec(limitsStr)[1]) * 1000,
@@ -30,9 +27,7 @@ export class TophProblemParser extends Parser {
       const [, amount, unit] = /, ([0-9.]+) (.*)/.exec(limitsStr);
       task.setMemoryLimit(parseFloat(amount) * (unit === 'MB' ? 1 : 1024));
 
-      const table = elem.querySelector(
-        '.problem-statement-portlet table:last-child',
-      );
+      const table = elem.querySelector('.problem-statement table:last-child');
 
       if (table !== null) {
         table.querySelectorAll('tbody tr').forEach(row => {